integrated test file into PyWren
