+++ pywren_ibm_cloud/__init__.py
@@ -17,5 +17,6 @@
 from pywren_ibm_cloud.wren import ibm_cf_executor
 from pywren_ibm_cloud import runtime
 from pywren_ibm_cloud.version import __version__
+from pywren_ibm_cloud.wren import test
 
 name = "pywren_ibm_cloud"
