--- test/data
@@ -1,5 +1,0 @@
-http://archive.ics.uci.edu/ml/machine-learning-databases/bag-of-words/vocab.enron.txt
-http://archive.ics.uci.edu/ml/machine-learning-databases/bag-of-words/vocab.kos.txt
-http://archive.ics.uci.edu/ml/machine-learning-databases/bag-of-words/vocab.nips.txt
-http://archive.ics.uci.edu/ml/machine-learning-databases/bag-of-words/vocab.nytimes.txt
-http://archive.ics.uci.edu/ml/machine-learning-databases/bag-of-words/vocab.pubmed.txt
